backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-MySQL==1.5.2
Flask-SQLAlchemy==3.1.1
openai==0.27.8
python-dotenv==1.0.0
PyMySQL==1.1.0